+++ formio/controllers/public.py
@@ -77,7 +77,7 @@ class FormioPublicController(http.Controller):
             'submission_date': fields.Datetime.now(),
         }
 
-        if post['data'].get('saveDraft') and not post['data'].get('submit'):
+        if post.get('saveDraft') or (post['data'].get('saveDraft') and not post['data'].get('submit')):
             vals['state'] = FORM_STATE_DRAFT
         else:
             vals['state'] = FORM_STATE_COMPLETE
@@ -144,10 +144,12 @@ class FormioPublicController(http.Controller):
             'submission_date': fields.Datetime.now(),
         }
 
-        if post['data'].get('saveDraft') and not post['data'].get('submit'):
+        if post.get('saveDraft') or (post['data'].get('saveDraft') and not post['data'].get('submit')):
             vals['state'] = FORM_STATE_DRAFT
+            vals['public_share'] = True
         else:
             vals['state'] = FORM_STATE_COMPLETE
+            vals['public_share'] = False
 
         context = {'tracking_disable': True}